backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.125.0",
    "pydantic>=2.12.5",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = "."
